Remove old OAuth tests based on a real world server
